스프링 카페 3단계 피드백 반영(#PR584) (#656)
* refactor(스프링 카페 2-3 단계 피드백(#PR584)): 네이밍 관련 변경

- 일급 컬렉션의 get method 네이밍 변경

* refactor(스프링 카페 2-3 단계 피드백(#PR584)): 각 도메인에 대한 Controller 분리

- Article과 Reply에 대한 Controller 분리

* refactor(스프링 카페 2-3 단계 피드백(#PR584)): 일급 컬렉션을 이용한 응집도 향상

- ReplyList에서 list를 꺼내는 것이 아닌 해당 class에서 규칙 확인

* refactor(스프링 카페 2-3 단계 피드백(#PR584)): Request DTO에 Build pattern 적용

- WriteRequest에 Build pattern 적용
- ReplyWriteRequest에 Build pattern 적용

* style(스프링 카페 2-3 단계 피드백(#PR584)): Article 관련 DTO 네이밍 변경

* style(스프링 카페 2-3 단계 피드백(#PR584)): Article DTO에 Build Pattern 적용

- UpdateArticleRequest에 Build pattern 적용

* refactor(스프링 카페 2-3 단계 피드백(#PR584)): 도메인 규칙 확인 로직 이동

기존 CheckPermissionInterceptor에서 하던 check logic을 모두 각각의 service layer로 이동

* refactor(스프링 카페 2-3 단계 피드백(#PR584)): 기능에 따른 Controller 분리

1. 댓글 출력 : ArticlePrintController -> ReplyPrintController 이동
2. 댓글 삭제 : ArticleDeleteController -> ReplyDeleteController 이동

* refactor(스프링 카페 2-3 단계 피드백(#PR584)): preHandler를 통해 sessionedUser 얻도록 변경

* style(스프링 카페 2-3 단계 피드백(#PR584)): 상수화된 쿼리 네이밍 변경

* style(스프링 카페 2-3 단계 피드백(#PR584)): domain configuration 파일 이름 변경

* style(스프링 카페 2-3 단계 피드백(#PR584)): 각 Repository에서 사용되는 Mapper를 spring bean으로 생성

* fix(스프링 카페 2-3 단계 피드백(#PR584)): url 변경

1. url 통일
2. bug fix

* refactor(스프링 카페 2-3 단계 피드백(#PR584)): 도전 과제 - ArticleDetail 정의

1. 게시글 상세보기를 하면 Service에서 새로 정의한 ArticleDetail 객체를 전달받는다.
2. 게시글 삭제하기를 하면 Service에서 ArticleDetail을 먼저 받은 후 삭제 가능한지 체크한다.

+++ src/main/java/com/kakao/cafe/application/reply/service/DeleteReplyService.java
@@ -2,6 +2,8 @@ package com.kakao.cafe.application.reply.service;
 
 import com.kakao.cafe.application.reply.port.in.DeleteReplyUseCase;
 import com.kakao.cafe.application.reply.port.out.DeleteReplyPort;
+import com.kakao.cafe.application.user.dto.UserInfo;
+import com.kakao.cafe.domain.user.exceptions.UnauthenticatedUserException;
 
 public class DeleteReplyService implements DeleteReplyUseCase {
 
@@ -12,12 +14,19 @@ public class DeleteReplyService implements DeleteReplyUseCase {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id, String userIdDeletingReply, UserInfo sessionedUser) throws UnauthenticatedUserException {
+        if (!sessionedUser.getUserId().equals(userIdDeletingReply)) {
+            throw new UnauthenticatedUserException("인증 오류");
+        }
         deleteReplyPort.delete(id);
     }
 
     @Override
-    public void deleteAllRepliesInArticle(int articleId) {
+    public void deleteAllRepliesInArticle(int articleId, String userIdDeletingReplies, UserInfo sessionedUser)
+        throws UnauthenticatedUserException {
+        if (!sessionedUser.getUserId().equals(userIdDeletingReplies)) {
+            throw new UnauthenticatedUserException("인증 오류");
+        }
         deleteReplyPort.deleteAllRepliesInArticle(articleId);
     }
 }
